requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
facebook-py==0.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.2.0
PySocks==1.7.1
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.13.2
siaskynet==2.2.0
six==1.15.0
style==1.1.0
telegram==0.0.1
tornado==6.1
tweepy==3.10.0
tzlocal==2.1
update==0.0.1
urllib3==1.26.4
Werkzeug==1.0.1
